requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
motor==3.6.0
pymongo==4.9.0
pydantic>=2.4.2,<3.0
pydantic-settings==2.1.0
python-multipart==0.0.6
paho-mqtt==1.6.1